enable auto launch on login

+++ mimiqHelper/AppDelegate.swift
@@ -8,44 +8,33 @@
 
 import Cocoa
 
-extension Notification.Name {
-    static let killLauncher = Notification.Name("killLauncher")
-}
-
-@NSApplicationMain
 class AppDelegate: NSObject, NSApplicationDelegate {
     func applicationDidFinishLaunching(_ aNotification: Notification) {
-        let mainAppIdentifier = "com.tiborbodecs.MainApplication"
-        let runningApps = NSWorkspace.shared.runningApplications
-        let isRunning = !runningApps.filter { $0.bundleIdentifier == mainAppIdentifier }.isEmpty
-
-        if !isRunning {
-            DistributedNotificationCenter.default().addObserver(self, selector: #selector(self.terminate), name: .killLauncher, object: mainAppIdentifier)
-
-            let path = Bundle.main.bundlePath as NSString
-            var components = path.pathComponents
-            components.removeLast()
-            components.removeLast()
-            components.removeLast()
-            components.append("MacOS")
-            components.append("MainApplication") //main app name
-
-            let newPath = NSString.path(withComponents: components)
-
-            NSWorkspace.shared.launchApplication(newPath)
+        let mainAppIdentifier = "com.wendyliga.mimiq"
+        
+        // Ensure the app is not already running
+        guard NSRunningApplication.runningApplications(withBundleIdentifier: mainAppIdentifier).isEmpty else {
+            NSApp.terminate(nil)
+            return
         }
-        else {
-            self.terminate()
-        }
+        
+        let path = Bundle.main.bundlePath as NSString
+        var components = path.pathComponents
+        components.removeLast()
+        components.removeLast()
+        components.removeLast()
+        components.append("MacOS")
+        components.append("mimiq")
+
+        let newPath = NSString.path(withComponents: components)
+
+        NSWorkspace.shared.launchApplication(newPath)
+        NSApp.terminate(nil)
+
     }
 
     func applicationWillTerminate(_ aNotification: Notification) {
         // Insert code here to tear down your application
     }
-
-    @objc func terminate() {
-        NSApp.terminate(nil)
-    }
-
 }
